feat(api): linking flow support
Before we supported linking flow we expected that received message
has "success" type, but it might be also "oidc", "form" or another
message type which requires additional login action.
This is also related to loginSetup and registerSetup functions
which may also return a "oidc", "form" or another message type.

IK-2288