feat: add render function
A new IKUICore.render function is added which is a replacement for
IKUICore.renderLogin and IKUICore.renderRegister functions. These
two functions are still present for backward compatibility.

This replacement is made because you can create and update your
authentication flow as you wish and you should not need to call
the IKUICore function according to what is the first node in your
flow.