feat: refresh ui container after an action click
The previous behaviour was that after you clicked an action you were
redirected to some different endpoint. You could choose from login,
forgotten and register actions. Other actions would not work.

Now, when you click on an action a new request will be sent to the
backend asking what is the next node which should be rendered.
If the action node has no outgoing connection, then we use the
previous behavior, which is deprecated now.