fix: support redirection in actions
If you want to redirect your page on an action click, you can provide
a path and the UI SDK will redirect you to that page. This operation
will not start the authentication flow from the beginning, but it
will continue in the existing one.

You can use one the these action names (login, forgotten and
register) and if you provide a path in the IKUICore.render call
(loginPath, registrationPath or forgotPasswordPath), you will be
redirected to the specified path. Anyway, you need to call on that
endpoint either IKUICore.render or IKUICore.renderForgotPasswordForm
function. These two functions will take the last response object
and will continue in the authentication flow.